feat: RestaurantReviewStat을 제거하고 Restaurant 엔티티로 통일
